--- ProcessingData/.idea/artifacts/ProcessingData_jar.xml
@@ -1,13 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="ProcessingData:jar">
-    <output-path>$PROJECT_DIR$/out/artifacts/ProcessingData_jar</output-path>
-    <root id="archive" name="ProcessingData.jar">
-      <element id="directory" name="META-INF">
-        <element id="file-copy" path="$PROJECT_DIR$/META-INF/MANIFEST.MF" />
-      </element>
-      <element id="module-output" name="ProcessingData" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.11.0/gson-2.11.0.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$MAVEN_REPOSITORY$/com/google/errorprone/error_prone_annotations/2.27.0/error_prone_annotations-2.27.0.jar" path-in-jar="/" />
-    </root>
-  </artifact>
-</component>
