--- ProcessingData/src/main/java/function/StatisticFunction.java
@@ -1,10 +1,0 @@
-package function;
-
-import model.JsonItem;
-
-import java.io.IOException;
-import java.util.List;
-
-public interface StatisticFunction {
-    String execute(List<JsonItem> jsonData) throws IOException;
-}
