--- ProcessingData/src/main/java/parser/JsonParser.java
@@ -1,9 +1,0 @@
-package parser;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.List;
-
-public interface JsonParser<T> {
-    List<T> fromJson(String json) throws IOException;
-}
