--- ProcessingData/src/main/resources/plugin-descriptor.properties
@@ -1,6 +1,0 @@
-description=My plugin
-name=my-plugin
-version=1.0.0
-opensearch.version=2.16.0
-java.version=20
-classname=MyPlugin
