+++ opensearch-plugin-template-java-main/.gitattributes
@@ -1,0 +1,1 @@
+*.bat text eol=crlf
