+++ opensearch-plugin-template-java-main/.github/CODEOWNERS
@@ -1,0 +1,1 @@
+*   @AmiStrn @dblock @saratvemulapalli @reta
