+++ opensearch-plugin-template-java-main/.github/workflows/CI.yml
@@ -1,0 +1,37 @@
+name: Build and Test Plugin Template
+
+on:
+  push:
+    branches:
+      - "*"
+  pull_request:
+    branches:
+      - "*"
+
+jobs:
+  build:
+    strategy:
+      matrix:
+        os:
+          - ubuntu-latest
+          - macOS-latest
+          - windows-latest
+        java:
+          - 17
+          - 21
+    name: Build and Test Plugin Template
+    runs-on: ${{ matrix.os }}
+
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up JDK ${{ matrix.java }}
+      uses: actions/setup-java@v4
+      with:
+        java-version: ${{ matrix.java }}
+        distribution: temurin
+    - name: Build and Run Tests
+      run: |
+        ./gradlew build
+    - name: Publish plugin ZIP to staging repository
+      run: |
+        ./gradlew publishPluginZipPublicationToZipStagingRepository
