+++ opensearch-plugin-template-java-main/MAINTAINERS.md
@@ -1,0 +1,12 @@
+## Overview
+
+This document contains a list of maintainers in this repo. See [opensearch-project/.github/RESPONSIBILITIES.md](https://github.com/opensearch-project/.github/blob/main/RESPONSIBILITIES.md#maintainer-responsibilities) that explains what the role of maintainer means, what maintainers do in this and other repos, and how they should be doing it. If you're interested in contributing, and becoming a maintainer, see [CONTRIBUTING](CONTRIBUTING.md).
+
+## Current Maintainers
+
+| Maintainer               | GitHub ID                                               | Affiliation |
+| ------------------------ | ------------------------------------------------------- | ----------- |
+| Amitai Stern             | [AmiStrn](https://github.com/AmiStrn)                   | Logz.io     |
+| Daniel "dB." Doubrovkine | [dblock](https://github.com/dblock)                     | Amazon      |
+| Sarat Vemulapalli        | [saratvemulapalli](https://github.com/saratvemulapalli) | Amazon      |
+| Andriy Redko             | [reta](https://github.com/reta)                         | Aiven       |
