+++ opensearch-plugin-template-java-main/NOTICE.txt
@@ -1,0 +1,2 @@
+OpenSearch (https://opensearch.org/)
+Copyright OpenSearch Contributors
