+++ opensearch-plugin-template-java-main/gradle.properties
@@ -1,0 +1,11 @@
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+# The OpenSearch Contributors require contributions made to
+# this file be licensed under the Apache-2.0 license or a
+# compatible open source license.
+#
+
+org.gradle.caching=true
+org.gradle.warning.mode=none
+org.gradle.parallel=true
